% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{download_dataset}
\alias{download_dataset}
\title{Download dataset}
\usage{
download_dataset(dataset_name)
}
\arguments{
\item{dataset_name}{string giving the name of the dataset to download}
}
\value{
if the dataset is stored as an R object (i.e., a \code{sceptre_object} or a list), then the dataset itself is returned. If, on the other hand, the dataset is stored as a collection of files, then a file path to the location of the dataset on disk is returned.
}
\description{
Downloads a dataset from \code{ExperimentHub()} to the user's machine. The available datasets are as follows:
\itemize{
\item "highmoi_example_mtx": Example high multiplicity-of-infection CRISPRi screen of putative enhancers (cell ranger format)
\item "highmoi_example_r_list": Example high multiplicity-of-infection CRISPRi screen of putative enhancers (R object format)
\item "lowmoi_example_r_list": Example low multiplicity-of-infection CRISPRko screen of gene transcription start sites
\item "parse_example_mtx": Example Parse Biosciences single-cell CRISPR screen
}
}
\examples{
data_dir <- download_dataset("highmoi_example_mtx")
highmoi_example_r_list <- download_dataset("highmoi_example_r_list")
}
